pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "faiss-storage-lib"
version = "0.1.0"
description = "Minimal FAISS-backed storage library for vectors and payloads."
readme = "README.md"
requires-python = ">=3.9"
dependencies = ["faiss-cpu", "numpy"]

[tool.setuptools.packages.find]
where = ["."]
include = ["faiss_storage_lib*"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "B"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["N999"]
